use i64 instead of i32 to avoid size issues

+++ src/version.rs
@@ -14,9 +14,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Default, Clone, Debug)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct Version {
-    pub major: i32,
-    pub minor: i32,
-    pub patch: i32,
+    pub major: i64,
+    pub minor: i64,
+    pub patch: i64,
     pub prerelease: Option<Vec<String>>,
 
     any: bool,
@@ -82,7 +82,7 @@ impl<'p> Version {
     }
 
     /// Constructs a version from its already parsed parts, e.g. `Version::from_parts(1, 2, 3, None)`.
-    pub fn from_parts(major: i32, minor: i32, patch: i32, prerelease: Option<String>) -> Self {
+    pub fn from_parts(major: i64, minor: i64, patch: i64, prerelease: Option<String>) -> Self {
         let prerelease = match prerelease {
             Some(pre) => {
                 let split: Vec<&str> = pre.split(".").collect();
